롱타입 버그 수정

+++ src/main/java/com/sellermatch/process/apply/domain/Apply.java
@@ -56,7 +56,7 @@ public class Apply {
     @Transient
     private String profileBizSort;
     @Transient
-    private Integer profileVolume;
+    private Long profileVolume;
     @Transient
     private String projMemId;
     @Transient
